chore: refactor urdf wheel tagging and handling
