+++ fission/src/urdf/URDFUserData.ts
@@ -1,2 +1,1 @@
-export const URDF_WHEEL_SOURCE_KEY = "wheelSource"
-export const URDF_AUTO_WHEEL_SOURCE = "urdf-auto"
+export const URDF_WHEEL_TAG = "urdfWheel"
