+++ fission/src/urdf/WheelDetector.ts
@@ -1,5 +1,5 @@
 import { mirabuf } from "@/proto/mirabuf"
-import { URDF_AUTO_WHEEL_SOURCE, URDF_WHEEL_SOURCE_KEY } from "./URDFUserData"
+import { URDF_WHEEL_TAG } from "./URDFUserData"
 
 const AXIS_PARALLEL_COS = 0.99 // axes must be this parallel to be the same axle direction
 const AXLE_ALIGN_COS = 0.98 // wheel-to-wheel displacement must align with the axis
@@ -163,7 +163,10 @@ export function detectAndTagWheels(assembly: mirabuf.Assembly): void {
         if (collinear.length > selected.length) selected = collinear.map(local => group.indices[local])
     }
 
-    if (selected.length === 0) return
+    if (selected.length === 0) {
+        console.error("No drivetrain found. Wheels will not be auto-assigned")
+        return
+    }
 
     for (const idx of selected) {
         for (const token of [candidates[pairs[idx].a].token, candidates[pairs[idx].b].token]) {
@@ -173,7 +176,7 @@ export function detectAndTagWheels(assembly: mirabuf.Assembly): void {
             jDef.userData.data ??= {}
             jDef.userData.data["wheel"] = "true"
             jDef.userData.data["wheelType"] = "0"
-            jDef.userData.data[URDF_WHEEL_SOURCE_KEY] = URDF_AUTO_WHEEL_SOURCE
+            jDef.userData.data[URDF_WHEEL_TAG] = "true"
         }
     }
 }
